Use react navigation header for ObsEdit
